requirements.txt
colorlog>=4.7.2
PyYAML>=5.4.1